Synchronize fbr_free to avoid use after free race
As soon as we marked the pool inactive it was possible for a thread to
free it instantly. This change puts the burden of freeing the pool's
primary allocation onto fbr_free instead of the last thread to exit the
pool. This is good for two reasons:
1. Use after free isn't possible. The primary allocation is available
   until the end of fbr_free's call.
2. Freeing is now a sync process. This means the call to fbr_free will
   not return until all the pool's resources have actually been freed.

+++ src/fbr_bm_alloc.h
@@ -23,10 +23,9 @@ struct fbr_bm_alloc_iterator {
 };
 
 /* I'm not forking ck to implement one silly function */
-CK_CC_INLINE static bool
-ck_bitmap_next_unset(const struct ck_bitmap *bitmap,
-		     struct ck_bitmap_iterator *i,
-		     unsigned int *bit)
+CK_CC_INLINE static bool ck_bitmap_next_unset(const struct ck_bitmap *bitmap,
+					      struct ck_bitmap_iterator *i,
+					      unsigned int *bit)
 {
 	unsigned int cache = i->cache;
 	unsigned int n_block = i->n_block;
@@ -48,7 +47,8 @@ ck_bitmap_next_unset(const struct ck_bitmap *bitmap,
 	}
 
 not_max:
-	*bit = CK_BITMAP_BLOCK * n_block + (unsigned int)ck_cc_ctz(~cache);
+	*bit = (unsigned int)CK_BITMAP_BLOCK * n_block +
+	       (unsigned int)ck_cc_ctz(~cache);
 	i->cache = cache | (cache + 1);
 	i->n_block = n_block;
 	return true;
